Rename packages to have a better hierarchy position -> position_conversion. Because that's what it does. AST <-> protocol processing -> ast_processing. Remove all mentions of non AST thing, and stop importing the position package
