+++ pkg/ast_processing/object.go
@@ -2,14 +2,14 @@ package processing
 
 import (
 	"github.com/google/go-jsonnet/ast"
-	"github.com/grafana/jsonnet-language-server/pkg/position"
 )
 
 // filterSelfScope takes in an array of objects (blocks delimited by curly braces) and
-//   returns a new array of objects, where only objects in scope of the first one are kept
+// returns a new array of objects, where only objects in scope of the first one are kept.
+
 // This is done by comparing the location ranges. If the range of the first object is
-//   contained within the range of another object, the latter object is removed because
-//   it is a parent of the first object.
+// contained within the range of another object, the latter object is removed because
+// it is a parent of the first object.
 func filterSelfScope(objs []*ast.DesugaredObject) (result []*ast.DesugaredObject) {
 	if len(objs) == 0 {
 		return objs
@@ -23,7 +23,7 @@ func filterSelfScope(objs []*ast.DesugaredObject) (result []*ast.DesugaredObject
 	for i < len(result) {
 		obj := result[i]
 		// If the current object is contained within the top level object, remove it
-		if position.RangeGreaterOrEqual(obj.LocRange, topLevel.LocRange) {
+		if RangeGreaterOrEqual(obj.LocRange, topLevel.LocRange) {
 			result = append(result[:i], result[i+1:]...)
 			continue
 		}
