+++ pkg/ast_processing/top_level_objects.go
